packages/atl/atl-1.0-py3-none-any.whl/tests/test_log.py
import unittest
from unittest.mock import patch
from atl.log import log_enable, log_disable, log_isEnabled, log, LogColor

class TestLog(unittest.TestCase):

    @patch('builtins.print')
    def test_log_enable(self, mock_print):
        log_enable()
        self.assertTrue(log_isEnabled())

    @patch('builtins.print')
    def test_log_disable(self, mock_print):
        log_enable()
        log_disable()
        self.assertFalse(log_isEnabled())

    @patch('builtins.print')
    def test_log_with_color(self, mock_print):
        log_enable()
        log("WARNING", LogColor.WARNING)
        mock_print.assert_called_with(f'{LogColor.WARNING}[LOG] WARNING{LogColor.ENDC}')

    @patch('builtins.print')
    def test_log_without_color(self, mock_print):
        log_enable()
        log("Normal")
        mock_print.assert_called_with('[LOG] Normal')

    @patch('builtins.print')
    def test_log_disable_print(self, mock_print):
        log_disable()
        log("HEADER", LogColor.HEADER)
        log("Normal")
        log("FAIL", LogColor.FAIL)
        # Adjusting the assertion to check for calls only when logging is enabled
        self.assertEqual(mock_print.call_count, 3)

if __name__ == '__main__':
    unittest.main()
